abstracted post upload form action
- checks if it returns form data or action result
- adding post to group thread adds correct group ids (also queried)
