+++ src/routes/(app)/groups/[id=id]/+page.server.ts
@@ -1,6 +1,10 @@
 import { AxiosHandler, handleReaction } from '$lib/utils/axios-handler';
+import { getPostUploadForm } from '$lib/utils/helpers';
 import { CookieName, type Post, type Tag, type Thread } from '$lib/utils/types';
-import { type Actions, error, fail, redirect } from '@sveltejs/kit';
+import {
+  type Actions, error, fail, isActionFailure,
+  redirect,
+} from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ cookies, params }) => {
@@ -37,35 +41,8 @@ export const load: PageServerLoad = async ({ cookies, params }) => {
 
 export const actions: Actions = {
   'add-post': async ({ request, params, cookies }) => {
-    const formData = await request.formData();
-    const files: File[] = [];
-    const content = `${formData.get('content')}`.replaceAll('\r\n\r\n', '\n');
-    const tagId = Number(formData.get('tag-id'));
-    let type = 'image';
-
-    if (formData.has('video')) {
-      const vid = formData.get('video') as File;
-      if (vid.size > 0) {
-        type = 'video';
-        files.push(vid);
-      }
-    } else if (formData.has('images')) {
-      files.push(...(formData.getAll('images') as File[]).filter((f) => f.size > 0));
-    }
-
-    if (files.length <= 0 && content.length <= 0) {
-      return fail(400, {
-        success: false,
-        message: 'Content must not be empty if no image or video uploaded.',
-      });
-    }
-
-    const form = new FormData();
-    form.append('type', type);
-    form.append('content', content);
-    form.append('approved', `${formData.has('group-approved')}`);
-    if (tagId > 0) form.append('tagId', `${tagId}`);
-    if (files.length > 0) files.forEach((f) => form.append('files', f));
+    const form = await getPostUploadForm(request);
+    if (isActionFailure(form)) return form;
 
     const res = await AxiosHandler.post(
       `/group/${params.id}/post`,
