+++ src/routes/(app)/groups/[id=id]/thread/[threadId=id]/+page.svelte
@@ -124,6 +124,8 @@
         formaction="?/add-post"
         placeholder="Add new post"
         postCallback={() => (showPostModal = false)}
+        groupId={data.group.id}
+        groupApproved
       />
     </ModalBody>
   </Modal>
