+++ src/routes/(app)/thread/[id=id]/+page.server.ts
@@ -1,6 +1,10 @@
 import { AxiosHandler, handleReaction } from '$lib/utils/axios-handler';
+import { getPostUploadForm } from '$lib/utils/helpers';
 import { CookieName, type Post, type Thread } from '$lib/utils/types';
-import { type Actions, error, fail, redirect } from '@sveltejs/kit';
+import {
+  type Actions, error, fail, isActionFailure,
+  redirect,
+} from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ cookies, params }) => {
@@ -58,32 +62,8 @@ export const actions: Actions = {
     return { success: true, message: res.message };
   },
   'add-post': async ({ request, params, cookies }) => {
-    const formData = await request.formData();
-    const files: File[] = [];
-    const content = `${formData.get('content')}`.replaceAll('\r\n\r\n', '\n');
-    let type = 'image';
-
-    if (formData.has('video')) {
-      const vid = formData.get('video') as File;
-      if (vid.size > 0) {
-        type = 'video';
-        files.push(vid);
-      }
-    } else if (formData.has('images')) {
-      files.push(...(formData.getAll('images') as File[]).filter((f) => f.size > 0));
-    }
-
-    if (files.length <= 0 && content.length <= 0) {
-      return fail(400, {
-        success: false,
-        message: 'Content must not be empty if no image or video uploaded.',
-      });
-    }
-
-    const form = new FormData();
-    form.append('type', type);
-    form.append('content', content);
-    if (files.length > 0) files.forEach((f) => form.append('files', f));
+    const form = await getPostUploadForm(request);
+    if (isActionFailure(form)) return form;
 
     const res = await AxiosHandler.post(
       `/thread/${params.id}/post`,
